Add dummy FT.HYBRID cmd to test tag scoring

+++ src/query.h
@@ -112,7 +112,7 @@ void QAST_SetGlobalFilters(QueryAST *ast, const QAST_GlobalFilterOptions *option
  * @return an iterator.
  */
 IndexIterator *QAST_Iterate(QueryAST *ast, const RSSearchOptions *options,
-                            RedisSearchCtx *sctx, ConcurrentSearchCtx *conc, uint32_t reqflags, QueryError *status);
+                            RedisSearchCtx *sctx, ConcurrentSearchCtx *conc, uint64_t reqflags, QueryError *status);
 
 /**
  * Expand the query using a pre-registered expander. Query expansion possibly
